use a custom subclass of ActiveResource::Connection if possible
