+++ lib/shopify_api.rb
@@ -4,7 +4,6 @@ require 'active_resource'
 require 'active_support/core_ext/class/attribute_accessors'
 require 'digest/md5'
 require 'base64'
-require 'active_resource/connection_ext'
 require 'active_resource/detailed_log_subscriber'
 require 'shopify_api/limits'
 require 'shopify_api/json_format'
@@ -22,3 +21,11 @@ require 'shopify_api/metafields'
 require 'shopify_api/countable'
 require 'shopify_api/resources'
 require 'shopify_api/session'
+require 'shopify_api/connection'
+
+# ActiveResource::Base.connection_class was introduced in https://github.com/rails/activeresource/pull/222
+if ShopifyAPI::Base.respond_to?(:connection_class)
+  ShopifyAPI::Base.connection_class = ShopifyAPI::Connection
+else
+  require 'active_resource/connection_ext'
+end
